+++ src/core/index.ts
@@ -1,6 +1,6 @@
 export { ReadError } from './domain/repositories/ReadError'
 export { WriteError } from './domain/repositories/WriteError'
-export { ApiConfig } from './infra/repositories/ApiConfig'
+export { ApiConfig, DataverseApiAuthMechanism } from './infra/repositories/ApiConfig'
 export { DvObjectOwnerNode, DvObjectType } from './domain/models/DvObjectOwnerNode'
 export { PublicationStatus } from './domain/models/PublicationStatus'
 export { StorageDriver } from './domain/models/StorageDriver'